requirements.txt
absl-py==1.4.0
accelerate==0.22.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
altgraph==0.17.3
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
astroid==2.12.13
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.16.0
backcall==0.2.0
backports.strenum==1.3.1
beautifulsoup4==4.12.3
begins==0.9
bitsandbytes
black==23.9.1
bleach==6.1.0
blinker==1.8.2
bokeh==3.1.1
cachetools==5.3.0
certifi 
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.0.6
crcmod==1.7
cryptography==41.0.4
cycler==0.11.0
Cython==0.29.33
dashscope==1.22.2
dataclasses-json==0.6.7
datasets==2.19.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
distro==1.8.0
dnnlib-util==0.3.1
docker-pycreds==0.4.0
easycython==1.0.7
einops==0.6.0
exceptiongroup==1.1.3
executing==1.2.0
export==0.2.1
fairscale==0.4.13
fastapi==0.104.1
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpy==0.3.1
filelock==3.8.2
Fiona==1.9.4.post1
fire==0.5.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
func_timeout==4.3.5
future==0.18.2
fvcore==0.1.5.post20221213
GDAL 
geojson==2.5.0
geopandas==0.13.0
getmac==0.8.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==0.4.6
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
griffe==0.48.0
grpcio==1.53.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.4
imageio==2.23.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==6.0.1
infinibatch==0.1.1
intel-openmp==2021.4.0
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.8.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
kornia==0.7.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy-object-proxy==1.8.0
lightning==2.1.0
lightning-utilities==0.9.0
linkify-it-py==2.0.2
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.22.0
mat4py==0.6.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mkl
mmcv-full 
mmengine==0.9.0
mmpretrain==1.1.0
model-index==0.1.11
modelindex==0.0.2
more-itertools==10.1.0
mpi4py==3.1.5
mpmath==1.3.0
mss
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1
nltk==3.8.1
notebook==7.3.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.0.2
openai==1.2.3
openai-clip==1.0.1
openai-whisper
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.27
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.1
panel==1.2.3
param==2.1.1
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pefile==2022.5.30
phx-class-registry==4.0.6
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
ply==3.11
portalocker==2.6.0
prettytable==3.9.0
prometheus_client==0.21.1
prompt-toolkit==3.0.36
protobuf==4.23.3
psutil
pudb==2023.1
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycocotools-windows 
pycparser==2.21
pycryptodome==3.16.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.14.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyiqa==0.1.6.4
pylint 
pyparsing==3.0.9
pyproj==3.4.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.2
pyqt6-tools==6.4.2.3.3
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==3.2.1
python-multipart==0.0.19
pytorch-lightning==1.4.2
pytz==2023.3.post1
pyviz_comms==3.0.4
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.2
pywinpty 
PyYAML==6.0
pyzmq==26.2.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
qudida==0.0.4
referencing==0.31.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.13.2
rsa==4.9
ruff==0.8.3
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.9.3
seaborn 
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==1.26.0
setproctitle==1.3.2
sgmllib3k==1.0.0
shapely==2.0.0
shellingham==1.5.4
sip==6.7.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.2
starlette==0.27.0
streamlit==1.40.1
sympy==1.12
tabulate==0.9.0
taming-transformers 
tb-nightly==2.14.0a20230521
tbb==2021.13.1
tenacity==8.5.0
tensorboard 
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.1
termcolor==2.1.1
terminado==0.18.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tiktoken==0.5.1
timeout-decorator==0.5.0
timm==0.6.12
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch 
torchaudio==2.3.1+cu118
torchmetrics==0.6.0
torchvision==0.18.1+cu118
tornado==6.4.2
tqdm==4.65.2
traitlets==5.8.1
transformers==4.46.3
ttach==0.0.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.13.0
uc-micro-py==1.0.2
ultralytics==8.0.120
uri-template==1.3.0
urllib3==1.26.20
urwid==2.2.2
urwid-readline==0.13
uvicorn==0.24.0.post1
wandb==0.15.4
watchdog==4.0.2
wcwidth==0.2.5
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.13
wincertstore==0.2
WMI==1.5.1
wrapt==1.14.1
xxhash==3.4.1
xyzservices==2025.1.0
yacs==0.1.8
yapf==0.32.0
yarl==1.9.2
zipp==3.12.0
